camss: csiphy: Make timer_clk_rate u32

+++ drivers/media/platform/qcom/camss-8x16/camss-csiphy.h
@@ -59,7 +59,7 @@ struct csiphy_device {
 	char irq_name[30];
 	struct camss_clock *clock;
 	int nclocks;
-	long timer_clk_rate;
+	u32 timer_clk_rate;
 	struct csiphy_config cfg;
 	struct v4l2_mbus_framefmt fmt[MSM_CSIPHY_PADS_NUM];
 };
